Fix a small bug in the Vista shortcut handling in the backupdb script where I forgot to make the restore shortcut target a script interpreter exe, so that the "Run as Administrator" bit is honoured.
Add a test for Vista for both backupdb and restoredb to see if the user is elevated and if not, request elevation by re-launching. This may not be ideal on non-Vista systems if the scripts are run by non-Admin users, so this is conditional on Vista and now a message is printed if run by a non-admin on pre-Vista systems.