In addition to the originally reported problem with being unable to save machine default configurations, there is another problem with the faulty way the GSS console UI handles these.
If the configuration server service does not receive a default user name from the client when initially creating the machine, the corresponding name it assigns for display in the console is just the machine name instead of the machine name and default user name combined. The initial configuration name happens to be the name as this, which is normally fine.

However, because the console incorrectly looks at machine default configurations as potentially clashing when saving a non-default named configuration, a default configuration which ended up being saved as a plain machine name will end up causing spurious conflicts for users which want to save an alternative named configuration with the machine name.

This extends the script slightly to ensure that machine default configurations that don't contain the substring " - ", and which are thus plain machine names, are also considered candidates to rename internally as potentially conflicting.